Align MCP DCR redirect policy with final redirect gate
